Bug fixes + layout changes
Fixed a bug with a newly added card;
This card had 2 jsonproperties with value "null". This caused and exception and would only allow 1814 cards to be read.
This is the only card that did something like that. So now I made the atk and def property nullable, just to be safe in the future.

Made some layout edits to the DetailPage.